requirements.txt
pandas
numpy
scikit-image
pillow
opencv-python
scipy
requests
google-generativeai
anthropic